convert to int

+++ src/com/polytechtours/events/Event.java
@@ -36,12 +36,12 @@ public abstract class Event implements Comparable<Event> {
 		this.date = date;
 	}
 
-	public double convertHourToSecond(double heure) {
-		return heure * 3600;
+	public int convertHourToSecond(double heure) {
+		return (int) (heure * 3600);
 	}
 
-	public double convertMinuteToSecond(double minute) {
-		return minute * 60;
+	public int convertMinuteToSecond(double minute) {
+		return (int)(minute * 60);
 	}
 	/**
 	 * @author Amine
